refactor(domains): Rename result.ts to confirm.ts for clarity
The file only contains daily confirmation logic (confirmDay,
autoConfirmRecentDays), so the name 'confirm.ts' better reflects
its purpose.
